refactor: rename to useDataTableContext
