+++ src/components/DataTable/TableLoadingComponent.tsx
@@ -1,4 +1,4 @@
-import { useDataTableContext } from "./useDataTable";
+import { useDataTableContext } from "./useDataTableContext";
 
 export interface TableLoadingComponentProps {
   render: (loading: boolean) => JSX.Element;
