+++ src/index.tsx
@@ -95,6 +95,6 @@ export * from "./components/DataTable/TableSorter";
 export * from "./components/DataTable/TableViewer";
 export * from "./components/DataTable/TextCell";
 export * from "./components/DataTable/useDataFromUrl";
-export * from "./components/DataTable/useDataTable";
+export * from "./components/DataTable/useDataTableContext";
 export * from "./components/Filter/FilterOptions";
 export * from "./components/Filter/GlobalFilter";
